Nuke YARG.Core builder and bring it in as a Unity package instead Also update NugetForUnity in the process
