+++ .gitignore
@@ -97,6 +97,9 @@ Assets/YargInput.cs
 
 # Rider
 .idea/
+
+# Manually-imported input packages
+# For users who have trouble with the package manager
 Packages/com.thenathannator.hidrogen/
 Packages/com.thenathannator.plasticband/
 
@@ -104,8 +107,13 @@ Packages/com.thenathannator.plasticband/
 Assets/Nyx
 Assets/Settings/TestPlayInfo.asset
 Assets/Settings/TestPlayInfo.asset.meta
-Assets/Plugins/YARG.Core/
-Assets/Plugins/YARG.Core.meta
 
 # Ignore .png~ files generated by Krita
 *.png~
+
+# Reload lock for submodule automation
+reload_lock~
+reload_lock~.meta
+# Submodule commit hash
+*.githash
+*.githash.meta
