+++ Assets/NuGet.config
@@ -1,15 +1,18 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <configuration>
-    <packageSources>
-       <clear/>
-       <add key="NuGet" value="http://www.nuget.org/api/v2/" />
-    </packageSources>
-    <disabledPackageSources />
-    <activePackageSource>
-       <add key="All" value="(Aggregate source)" />
-    </activePackageSource>
-    <config>
-       <add key="repositoryPath" value="./Packages" />
-       <add key="DefaultPushSource" value="http://www.nuget.org/api/v2/" />
-    </config>
+  <packageSources>
+    <add key="NuGet" value="http://www.nuget.org/api/v2/" />
+  </packageSources>
+  <disabledPackageSources />
+  <packageSourceCredentials />
+  <activePackageSource>
+    <add key="All" value="(Aggregate source)" />
+  </activePackageSource>
+  <config>
+    <add key="packageInstallLocation" value="CustomWithinAssets" />
+    <add key="repositoryPath" value="./Packages" />
+    <add key="PackagesConfigDirectoryPath" value="." />
+    <add key="DefaultPushSource" value="http://www.nuget.org/api/v2/" />
+    <add key="slimRestore" value="true" />
+  </config>
 </configuration>
