+++ Assets/Plugins/Editor/Submodule.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3e48859e7697fd448b1d86f072c33a9f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
