+++ Assets/Plugins/Editor/Submodule/ProjectAdder.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6b7d813cabf71054c8545c9c958c0f4a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
