+++ Assets/Plugins/Editor/Submodule/SubmoduleHelper.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: f686e3504a96e5b4b82641647a5655b0
+guid: a3e9f8a1791f10643bc394210140ac0a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
