+++ Assets/Plugins/Editor/Submodule/YARG.Editor.Submodules.asmdef
@@ -1,0 +1,16 @@
+{
+    "name": "YARG.Editor.Submodules",
+    "rootNamespace": "",
+    "references": [
+        "GUID:d9149a4e0ef435c45b10905a79471258"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": true,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
