+++ Assets/Plugins/Editor/Submodule/YARG.Editor.Submodules.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 00fc6afd415b19242b970f3e0b794b69
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
