+++ Assets/packages.config
@@ -1,15 +1,15 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="FuzzySharp" version="2.0.2" />
-  <package id="ManagedBass" version="3.1.1" />
-  <package id="ManagedBass.Fx" version="3.1.1" />
-  <package id="ManagedBass.Mix" version="3.1.1" />
-  <package id="Melanchall.DryWetMidi.Nativeless" version="7.0.0" />
+  <package id="FuzzySharp" version="2.0.2" manuallyInstalled="true" />
+  <package id="ManagedBass" version="3.1.1" manuallyInstalled="true" />
+  <package id="ManagedBass.Fx" version="3.1.1" manuallyInstalled="true" />
+  <package id="ManagedBass.Mix" version="3.1.1" manuallyInstalled="true" />
+  <package id="Melanchall.DryWetMidi.Nativeless" version="7.0.0" manuallyInstalled="true" />
   <package id="Microsoft.NETCore.Platforms" version="7.0.1" />
   <package id="NETStandard.Library" version="2.0.3" />
-  <package id="sqlite-net" version="1.6.292" />
-  <package id="System.Diagnostics.Tracing" version="4.3.0" />
-  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" />
-  <package id="UniRx" version="5.4.1.0" />
-  <package id="ZString" version="2.5.1" />
+  <package id="System.Diagnostics.Tracing" version="4.3.0" manuallyInstalled="true" />
+  <package id="System.Runtime.CompilerServices.Unsafe" version="6.0.0" manuallyInstalled="true" />
+  <package id="UniRx" version="5.4.1" manuallyInstalled="true" />
+  <package id="ZString" version="2.5.1" manuallyInstalled="true" />
+  <package id="sqlite-net" version="1.6.292" manuallyInstalled="true" />
 </packages>
