+++ Packages/packages-lock.json
@@ -7,6 +7,13 @@
       "dependencies": {},
       "url": "https://package.openupm.com"
     },
+    "com.cysharp.csprojmodifier": {
+      "version": "https://github.com/Cysharp/CsprojModifier.git?path=src/CsprojModifier/Assets/CsprojModifier",
+      "depth": 0,
+      "source": "git",
+      "dependencies": {},
+      "hash": "e82080dac67f22fa0ef0b7c0aceadda92df76ec3"
+    },
     "com.cysharp.unitask": {
       "version": "2.5.3",
       "depth": 0,
@@ -15,7 +22,7 @@
       "url": "https://package.openupm.com"
     },
     "com.github-glitchenzo.nugetforunity": {
-      "version": "3.1.3",
+      "version": "4.5.0",
       "depth": 0,
       "source": "registry",
       "dependencies": {},
@@ -303,6 +310,12 @@
       "source": "git",
       "dependencies": {},
       "hash": "71879e71aa72465a79bdf009d7d71f9f7ab83066"
+    },
+    "in.yarg.core": {
+      "version": "file:../YARG.Core/YARG.Core",
+      "depth": 0,
+      "source": "local",
+      "dependencies": {}
     },
     "jp.keijiro.minis": {
       "version": "https://github.com/TheNathannator/Minis.git?path=/Packages/jp.keijiro.minis#yarg",
