+++ ProjectSettings/CsprojModifierSettings.json
@@ -1,0 +1,1 @@
+{"AdditionalImports":[{"Path":"EnableNullability.props","Position":2}],"AdditionalImportsAdditionalProjects":["YARG.Core.Package.csproj"],"EnableAddAnalyzerReferences":false,"AddAnalyzerReferencesAdditionalProjects":[]}
